used CRTP for virtual detector
