+++ slsDetectorServers/matterhonServer/include/VirtualMatterhornServer.h
@@ -1,0 +1,26 @@
+
+#include "BaseMatterhornServer.h"
+
+namespace sls {
+
+class VirtualMatterhornServer
+    : public BaseMatterhornServer<VirtualMatterhornServer> {
+
+  public:
+    /**
+     * Constructor
+     * Starts up a virtual Matterhorn server.
+     * Assembles a virtual Matterhorn server using TCP and UDP detector
+     * interfaces throws an exception in case of failure
+     * @param port TCP/IP port number
+     */
+    explicit VirtualMatterhornServer(uint16_t port = DEFAULT_TCP_CNTRL_PORTNO);
+
+    ~VirtualMatterhornServer() = default;
+
+    ReturnCode initial_checks(ServerInterface &socket);
+
+    ReturnCode get_update_mode(ServerInterface &socket);
+};
+
+} // namespace sls
